requirements.txt
Django==2.2.3
Faker==2.0.0
djangorestframework==3.10.1